pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metabase-mcp-plus"
version = "1.0.2"
description = "Enhanced MCP server for Metabase — query databases, manage dashboards, cards, and collections from AI assistants like Claude and Cursor"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Sahil Chachan"},
]
keywords = [
    "metabase",
    "mcp",
    "model-context-protocol",
    "ai",
    "claude",
    "cursor",
    "anthropic",
    "fastmcp",
    "sql",
    "database",
    "analytics",
    "business-intelligence",
    "bi",
    "data-analysis",
    "ai-assistant",
    "llm",
    "chatgpt",
    "automation",
    "api",
    "python",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastmcp>=2.12.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pytest-cov>=7.0.0",
]

[project.scripts]
metabase-mcp-plus = "server:main"

[project.urls]
Homepage = "https://github.com/schachan/metabase-mcp"
Repository = "https://github.com/schachan/metabase-mcp"
Issues = "https://github.com/schachan/metabase-mcp/issues"

[tool.hatch.build]
exclude = [
    "tests/",
    ".github/",
    ".env*",
    ".gitignore",
    ".python-version",
    "uv.lock",
    "dist/",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space" 
